chore: move spawnProcess to new file
