+++ scripts/generate-clients/code-gen.js
@@ -1,7 +1,7 @@
 const path = require("path");
 const { copyFileSync, emptyDirSync } = require("fs-extra");
 const { readdirSync, lstatSync } = require("fs");
-const { spawn } = require("child_process");
+const { spawnProcess } = require("./spawn-process");
 const { CODE_GEN_ROOT, CODE_GEN_INPUT_DIR } = require("./code-gen-dir");
 
 async function generateClients(models) {
@@ -34,24 +34,4 @@ async function generateClients(models) {
   );
 }
 
-const spawnProcess = (command, args = [], options = {}) =>
-  new Promise((resolve, reject) => {
-    try {
-      const ls = spawn(command, args, options);
-      ls.stdout.on("data", data => {
-        console.log(data.toString());
-      });
-      ls.stderr.on("data", data => {
-        console.error(`stderr: ${data.toString()}`);
-      });
-
-      ls.on("close", code => {
-        console.log(`child process exited with code ${code}`);
-        resolve();
-      });
-    } catch (e) {
-      reject(e);
-    }
-  });
-
-module.exports = { generateClients, spawnProcess };
+module.exports = { generateClients };
